django-admin/requirements.txt
Django==5.0.4
psycopg[binary]==3.1.18
python-dotenv==1.0.1
split-settings==1.0.0
uwsgi==2.0.26
python-telegram-bot==21.11.1
loguru==0.7.3